Add sprite info hook
